fix: Color contrast
The color contrast for showing PP was not accessible. Neither was the
color contrast for the footer link with a link to my github profile.
This has been fixed by using darker shades of the red/blue that were
being used. Now this passes the Accessibility Insights for Web Fast Pass