feat: Support playoff games
This adds support for playoff games. When a playoff game is shown, if
the series is tied, it will show "Series {home team wins}-{away team
wins}". If a team is leading the series, it will show what team is
leading. If a team has won the series, it will show what team won (e.g.
VAN wins 4-3).