fix: Show the series status from the api
This fixes an issue where the next round series shows as being Tied 4-4
because the "leagueRecord" field that comes back from the api actually
shows the full record of the entire playoffs (ie it does not reset for
each new round a team gets into). Unfortunately, there is not an easy
way to determine what the series score is without jumping through hoops,
so instead the seriesStatusShort was used as this shows a string like
"CAR Leads 3-2".