fix: fixing the team record and logo
Sometimes games do not return the ot field for the leaguerecord. When
this happens, just the wins and losses will be displayed.

For some pre-season games (e.g. global series), games are played with a
team not in the NHL. This means a logo does not exist for them.
Previously, it used to display the team id with an orange background.
Now it displays the team abbreviation and the height and width are
aligned with the rest of the logos.

Some other minor changes:

* tailwind config js has been run through prettier
* the PR gameType was added for pre season games.
* Some misc code re-org happened.