feat: support postponed games
Adds support for postponed games.
Also fixes the SO in regular season display (it was previously
displaying 2OT which doesnt make sense until playoffs).

Also changing the types around a bit to make things a little cleaner. A
separate component type layer was created so as not to couple too much
with the api types.