fix: standings page when a team has not played a game
When a team has yet to play a game for the season, their "streak" does
not exist in the standings api response. Previously, the code assumed it
was always present. This has been fixed to ensure the "streak" field is
now optional. Additionaly, a StreakCell component has been created to
accomodate this new requirement and shows 2 dashes ("--") when the
"streak" field is not present (this matches what nhl.com shows as well).

Additionally adding unit tests to cover these cases.