feat: add support for scoring type
Shows whether the goal was even, shg, or pp depending on how the goal
was scored.

Also changes the score display of the individual goal to prioritize the
team with the higher score. Previously, if the away team was in the
lead, it would display their score second (ie "0-1 VAN"). This has now
been fixed to display their score first (ie "1-0 VAN").