fix: client hydration errors
Fixing the following error:

> Hydration failed because the initial UI does not match what was rendered on the server.

This was due to the startTime being different on the server than on the
client since on the server it is in Eastern time but on the client it
uses the users local time zone (which may be different).

This is fixed by using a `useHydration` hook with a Suspense boundary.

More info here:

https://francoisbest.com/posts/2023/displaying-local-times-in-nextjs