fix: start time from "popping"
Since the start time from the server is in EST, if the user is in a
different time zone the start time "pops" (ie it shows up in est then
changes to their local time). To mitigate this, a time skeleton is shown
briefly until the client has been hydrated. Then the time can be
displayed in the users local time zone.